resourceservice/src/main/java/com/eldorado/resourceservice/service/UserDetailsServiceImpl.java
package com.eldorado.resourceservice.service;

import org.slf4j.Logger;
import org.slf4j.LoggerFactory;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.security.core.userdetails.UserDetails;
import org.springframework.security.core.userdetails.UserDetailsService;
import org.springframework.security.core.userdetails.UsernameNotFoundException;
import org.springframework.stereotype.Service;

import com.eldorado.resourceservice.models.AuthUser;
import com.eldorado.resourceservice.repository.UserRepository;

@Service
public class UserDetailsServiceImpl implements UserDetailsService {
	
	private static Logger logger = LoggerFactory.getLogger(UserDetailsServiceImpl.class);

	@Autowired
	private UserRepository userRepo;

	@Override
	public UserDetails loadUserByUsername(String userName) throws UsernameNotFoundException {

		AuthUser user = userRepo.findByEmail(userName);
		if (user == null) {
			logger.error("User is not present with username "+userName);
			throw new UsernameNotFoundException("User is not present !");
		}
		logger.error("User is present with username "+userName);
		return new org.springframework.security.core.userdetails.User(user.getEmail(), user.getPassword(),
				user.getRoles());
	}

}
